feat: [FEM-36] Update new transaction viewModel update logic

+++ source/app/src/main/java/com/charmflex/flexiexpensesmanager/features/account/data/responses/AccountSummaryResponse.kt
@@ -12,9 +12,9 @@ internal data class AccountSummaryResponse(
     @ColumnInfo("account_name")
     val accountName: String?,
     @ColumnInfo("out_amount")
-    val outAmount: Int,
+    val outAmount: Long,
     @ColumnInfo("in_amount")
-    val inAmount: Int,
+    val inAmount: Long,
     @ColumnInfo("account_initial_amount")
     val initialAmount: Int
 )
